+++ .gitignore
@@ -1,3 +1,4 @@
 .venv/
 .ipynb_checkpoints/
 __pycache__/
+*.egg-info/
