+++ requirements.txt
@@ -1,91 +1,46 @@
-aiohttp==3.8.1
-aiohttp-retry==2.5.2
-aiosignal==1.2.0
-amqp==5.1.1
-appdirs==1.4.4
-async-timeout==4.0.2
-asyncssh==2.11.0
-atpublic==3.0.1
-attrs==21.4.0
-billiard==4.0.0
-black==22.6.0
-celery==5.3.0a1
-certifi==2022.6.15
-cffi==1.15.1
-charset-normalizer==2.1.0
-click==8.1.3
-click-didyoumean==0.3.0
-click-plugins==1.1.1
-click-repl==0.2.0
-colorama==0.4.5
-commonmark==0.9.1
-configobj==5.0.6
-cryptography==37.0.4
-dictdiffer==0.9.0
-diskcache==5.4.0
-distro==1.7.0
-dpath==2.0.6
-dulwich==0.20.45
-dvc==2.15.0
-dvc-data==0.1.5
-dvc-objects==0.1.5
-dvc-render==0.0.6
-dvc-task==0.1.0
-dvclive==0.9.0
-flatten-dict==0.4.2
-flufl.lock==7.0
-frozenlist==1.3.0
-fsspec==2022.5.0
-ftfy==6.1.1
-funcy==1.17
-future==0.18.2
-gitdb==4.0.9
-GitPython==3.1.27
-grandalf==0.6
-idna==3.3
-kombu==5.3.0a1
-mailchecker==4.1.19
-multidict==6.0.2
-mypy-extensions==0.4.3
-nanotime==0.5.2
-networkx==2.8.5
+absl-py==1.2.0
+asttokens==2.0.8
+backcall==0.2.0
+cycler==0.11.0
+-e git+https://github.com/virati/DBS_Readouts.git@aacd8c172e033d2be50802fc231ec55b1b3b9426#egg=dbread
+debugpy==1.6.3
+decorator==5.1.1
+entrypoints==0.4
+etils==0.7.1
+executing==0.10.0
+fonttools==4.36.0
+importlib-resources==5.9.0
+ipykernel==6.15.1
+ipython==8.4.0
+jax==0.3.16
+jaxlib==0.3.15
+jedi==0.18.1
+jupyter-client==7.3.4
+jupyter-core==4.11.1
+kiwisolver==1.4.4
+matplotlib==3.5.3
+matplotlib-inline==0.1.6
+nest-asyncio==1.5.5
+numpy==1.23.2
+opt-einsum==3.3.0
 packaging==21.3
-pathspec==0.9.0
-phonenumbers==8.12.52
-platformdirs==2.5.2
+parso==0.8.3
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==9.2.0
 prompt-toolkit==3.0.30
 psutil==5.9.1
-pycparser==2.21
-pydot==1.4.2
-pygit2==1.10.0
-Pygments==2.12.0
-pygtrie==2.5.0
+ptyprocess==0.7.0
+pure-eval==0.2.2
+Pygments==2.13.0
 pyparsing==3.0.9
-python-benedict==0.25.2
 python-dateutil==2.8.2
-python-fsutil==0.6.1
-python-slugify==6.1.2
-pytz==2022.1
-PyYAML==6.0
-requests==2.28.1
-rich==12.5.1
-ruamel.yaml==0.17.21
-ruamel.yaml.clib==0.2.6
-scmrepo==0.0.25
-shortuuid==1.0.9
-shtab==1.5.5
+pyzmq==23.2.1
+scipy==1.9.0
 six==1.16.0
-smmap==5.0.0
-tabulate==0.8.10
-text-unidecode==1.3
-toml==0.10.2
-tomli==2.0.1
-tqdm==4.64.0
+stack-data==0.4.0
+tornado==6.2
+traitlets==5.3.0
 typing_extensions==4.3.0
-urllib3==1.26.11
-vine==5.0.0
-voluptuous==0.13.1
 wcwidth==0.2.5
-xmltodict==0.13.0
-yarl==1.7.2
-zc.lockfile==2.0
+zipp==3.8.1
